update model_state & add venv

+++ 0x0F-python-object_relational_mapping/model_state.py
@@ -6,12 +6,13 @@ from sqlalchemy.ext.declarative import declarative_base
 
 Base = declarative_base()
 
+
 class State(Base):
     """state table class"""
     __tablename__ = 'states'
-    id = Column("id", Integer, primary_key=True, nullable=False)
+    id = Column("id", Integer(11), primary_key=True, nullable=False)
     name = Column("name", String(128), nullable=False)
-    
+
     def __init__(self, id, name):
         self.id = id
         self.name = name
